[kernels] Improve agreement between results of `BosonCorr` and `BosonAutoCorr`
It looks like sampling is somewhat problematic when a major part of
spectral weight is localized near an edge of the energy window. In
particular, results of `BosonAutoCorr` (`energy_window.first == 0`)
systematically deviate from those of `BosonCorr` obtained from the same
symmetric input data.

In order to solve this problem, `BosonAutoCorr` kernels have been
reworked. The sought spectral function is now defined on the full energy
axis, while application of the kernels effectively extracts only the
symmetric component of the spectrum. `back_transform()` and `compute_tail()`
make this extraction explicit and symmetrize the spectrum as
$(A(e) + A(-e)) / 2$ before performing the back-transformation/tail
computation.

User API changes:

* Requested energy window must be centered around zero
  for `BosonAutoCorr`.
* From now on, `BosonCorr` and `BosonAutoCorr` expect the same solution
  norm $\pi\chi(i\Omega=0)$ (no need to divide this value by 2 for
  `BosonAutoCorr`).